refactor: split doordash/queries.ts into focused query modules
